Created a viewmodel to homeActivity to check the connectivity on the application. Working WELL! This permit the app to be able to check for internet connectivity on LIVEDATA... so when internet connection drops, the user will be prompted with a Snackbar saying the internet status... and when it comes back too.
Changed the API Level to 21 to use the Connectivity class found in an article about connectivity check on livedata. No changes on class.
Link located inside the class comments.
Renamed some classes to match a good practice and make the app more understandable.

ktlint and ktlint format to check for issues on code.